fix(run-engine): waitpoint update misleading error logs
The `completeWaitpoint` method ignores Prisma update errors of type `P2025` to
handle concurrent attempts to complete a waitpoint token. We have an
error hook on the Prisma client which logs these errors though.

This PR switches to `updateMany` which doesn't throw an error if the
waitpoint is in an unexpected state. The behavior remains otherwise
unchanged. We currently continue even if the waitpoint is not in `PENDING`,
which could lead to race conditions in certain corner cases, though very
unlikely. We added a log line to see how often concurrent completion
attempts happen.