chore(releases): automatically update the lockfile in changeset PRs
This is a step which we currently need to do manually and it's rather painful. The lockfile update is necessary due to cross references in our packages.

Added it as a separate job instead of a step to start from fresh workspace, as the state that the `changeset` step leaves the workdir is not explicitly clear to the reader.